requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.3.4
httpx==0.26.0
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
